Fix weather, maybe

+++ src/main/java/uk/co/ystv/ystvbot/commands/Weather.java
@@ -29,7 +29,7 @@ public class Weather extends Command {
 			Map<String, Object> wind = (Map<String, Object>) map.get("wind");
 			map = (Map<String, Object>) list.get(0);
 			String[] dirs = new String[] { "N", "NE", "E", "SE", "S", "SW", "W", "NW" };
-			String wind_direction = dirs[((Integer) wind.get("deg") + 22) / 45 % 8];
+			String wind_direction = dirs[(((Double) wind.get("deg")).intValue() + 22) / 45 % 8];
 
 			return map.get("main") + ", Wind: " + wind.get("speed") + "m/s (" + wind_direction + "), Temperature: " + main.get("temp") + "C (Min: " + main.get("temp_min") + "C, Max: " + main.get("temp_max") + "C)";
 		} catch (MalformedURLException e) {
